Working with rovers and quads in empty world
The simulation works with rovers and quads in an empty world
for circular trajectories.

There is one box on the ground for the rovers to avoid. It should
work with noise as well, although results will be messy for the
circular trajectories.

This commit fixed the naming issue with the rovers. Note that the
rover numbering starts after the last quad number; i.e. if there
are 5 quads and 2 rovers, the rovers will be named "rover5" and
"rover6".