Fixed IO_controller_R1 CBF formulation
The CBF formulation for the R1 rovers was incorrect when dealing with
quadrotors, since only the quadrotors' x and y positions were
taken into account. This has been fixed so that the rovers take
into account the z distance as well; e.g. the rovers will not
avoid quadrotors which are sufficiently high overhead.